Add caching to the tile overlay renderer
This should cut down on the number of render calls needed to render
Minecraft map tiles in the map views. If this behavior is not desired,
developers can set the ephemeral property on MinecraftMapView to revert
to the old behavior. This functionality is not available in SwiftUI, so
caching will always occur.