CBK-5: Split CubiomesKit
This is still an in-progress refactor, but this commit notably splits
the original CubiomesKit package in to CubiomesKitCore and
CubiomesMapKit. A conglomerate package 'CubiomesKit' also exists,
provided that certain modules can be imported.

The package itself builds fine under Linux, but extensive testing on the
macOS side has yet to be done, and the testing suite will need to be
overhauled to disable MapKit-specific tests.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>