Use Vapor's Swiftly action
The author of the Swiftly action being used to set up the tooling
on the CI front deleted their repo and most of their content to move
to Codeberg. To prevent CI from breaking for now while CI is still
being done on GitHub Actions, the action in question has been
replaced.

In the future, we should consider moving CI efforts off GitHub.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>